LATX, fix: Preserve TU jump state during SMC TB reload
Background/root cause:
SMC reload treated the aliased jmp_reset_offset fields of a TU-jump TB as generic jump slots and called tb_reset_jump(). tb_reset_jump() asserts !use_tu_jmp(tb), while TU-internal jumps must retain their TU state.

Keep generic jump reset limited to non-TU TBs. The focused regression invokes the production SMC reload function with a TU-jump TB and the production tb_reset_jump() implementation.

Test:
  RED: build-pr299-tdd/test-latx-smc-reload-tu
  Observed: exit 134 (SIGABRT): Assertion '!use_tu_jmp(tb)' failed.
  GREEN: meson test -C build-pr299-tdd --no-rebuild latx-smc-reload-tu
  Observed: 1/1 latx-smc-reload-tu OK.
