+++ meson.build
@@ -578,6 +578,23 @@ foreach target : target_dirs
                  build_by_default: false,
                  name_suffix: 'fa')
 
+  if target == 'x86_64-linux-user' and 'CONFIG_LATX_O1' in config_host
+    smc_reload_tu_test = executable('test-latx-smc-reload-tu',
+                                    files('target/i386/latx/sbt/tests/smc-reload-tu-test.c',
+                                          'target/i386/latx/sbt/smc_reload.c',
+                                          'target/i386/latx/sbt/aot_smc_reload.c',
+                                          'accel/tcg/tb-jump.c'),
+                                    c_args: c_args + ['-ffunction-sections',
+                                        '-fvisibility=hidden',
+                                        '-I' + (meson.current_build_dir() /
+                                        ('libqemu-' + target + '.fa.p'))],
+                                    dependencies: [glib],
+                                    include_directories: target_inc,
+                                    link_args: ['-Wl,--gc-sections',
+                                        '-Wl,--no-export-dynamic'])
+    test('latx-smc-reload-tu', smc_reload_tu_test)
+  endif
+
     execs = [{
       'name': 'latx-' + target_name,
       'gui': false,
